Module Jocalf__Interp

Main interpretation functions of interpreter.

val interp_expr : string -> string

interp_expr s interprets s as an expression. This is useful for testing in a test suite or in the toplevel

val interp_phrase : (string * Jocalf.Eval.env * Jocalf.Eval.state) -> string * Jocalf.Eval.env * Jocalf.Eval.state

inter_phrase (s, env, st) interprets s in an environment and state, yielding a new environment and state. The REPL uses this to interpret input from the user.